reuse functions from cudnn
